refactor(lifecycle): unify Unity editor state snapshots
Replace duplicated string-based lifecycle and generation fragments with shared typed snapshots across IPC, persistence, application projections, schemas, and tests.

Harden screenshot evidence capture against unstable presentation state and cleanup failures, and bind project identity at bootstrap and compile recovery boundaries.

+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Execution/PlanToken/Environment/PlanTokenEnvironmentSnapshot.cs
@@ -17,5 +17,5 @@ namespace MackySoft.Ucli.Unity.Execution.PlanToken
         string ProjectFingerprint,
         string UnityVersion,
         IpcCompileState CompileState,
-        int DomainReloadGeneration);
+        long DomainReloadGeneration);
 }
